feat: create category reducers and add data

+++ client/admin-dashboard/src/assets/redux/actions/auth.action.js
@@ -60,11 +60,22 @@ export const userLoggedin = () => {
 
 
 
-export const signout = ()=>{
-  return async dispatch=>{
-    localStorage.clear();
-    dispatch({
-      type:authConstants.LOGOUT_REQUEST
-    })
+export const signout = () => {
+  return async dispatch => {
+
+      dispatch({ type: authConstants.LOGOUT_REQUEST });
+      const res = await axios.post(`/admin/signout`);
+
+      if(res.status === 200){
+          localStorage.clear();
+          dispatch({ type: authConstants.LOGOUT_SUCCESS });
+      }else{
+          dispatch({
+              type: authConstants.LOGOUT_FAILURE,
+              payload: { error: res.data.error }
+          });
+      }
+
+      
   }
 }
